backend/requirements.txt
flask==2.0.1
gunicorn==21.2.0
google-cloud-documentai==2.20.1
flask-cors==4.0.0
werkzeug==2.0.1
pydantic==2.5.1
